requirements.txt
modal
fastapi
uvicorn
opencv-python-headless
pillow
numpy
torch
torchvision
basicsr
realesrgan